Components
1.Question 1
What characteristics distinguish a Solid State Drive from a Hard Disk Drive? Check all that apply.

<A>   Uses Disk Platters
<B>   High RPMs
<C>   Non-moving parts
<D>   Smaller form factor

AnS:  <C>   Non-moving parts


2.Question 2
True or false: If you plug in a 220v appliance into a 120v outlet, the appliance could get damaged.

<A>   TRUE
<B>   FALSE

AnS:<A>   TRUE


3.Question 3
How long will it take to transfer a file size of 1GB on a USB 2.0 and a USB 3.0 drive?

   USB 2.0  Transfer speeds of 480 mb/s
   USB 3.0  Transfer Speeds of 5 GB/s
   USB 3.1 Transfer Speeds of 10gb/s


<A>    ~20 seconds on a USB 2.0 drive; ~2 seconds on a USB 3.0 drive
<B>    ~17 seconds on a USB 2.0 drive; ~.02 seconds on a USB 3.0 drive
<C>    ~ 17 seconds on a USB 2.0 drive; ~2 seconds on a USB 3.0 drive
<D>    ~ 20 seconds on a USB 2.0 drive; ~.02 seconds on a USB 3.0 drive


AnS:<C>    ~ 17 seconds on a USB 2.0 drive; ~2 seconds on a USB 3.0 drive


4.Question 4
Which of these is used to charge devices? Check all that apply.

<A>   Picture of USB A.
<B>   Picture of USB C.
<C>   Picture of DisplayPort.
<D>   Picture of HDMI Port.


AnS:<B>   Picture of USB C.